+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 ## Version 1.0.6
 
-  * 
+  * Add badge\_links to Meeting
+  * Event.add\_column method gains a required attirbute (default false) to control whether parents are required to enter something
+  * Event::Column gains parent\_required attribute (default false)
+  * Event::Attendance gains get\_audit\_trail(api) method
 
 ## Version 1.0.5
 
-  * Fix undefined method youth_section? within Model.require\_ability\_to
+  * Fix undefined method youth\_section? within Model.require\_ability\_to
 
 ## Version 1.0.4